# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EJ Learning adaptive learning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content/materials/{materialId}/file:
    get:
      operationId: getMaterialFile
      tags: [content]
      summary: The approved file for a source material
      description: >
        Streams the stored file, currently always a PDF. Viewers accept a
        #page=N fragment, so a lesson can open the book at its own pages.
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: No approved version, or the stored file is missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /teacher/assessment-sheet:
    get:
      operationId: getAssessmentSheet
      tags: [teacher]
      summary: The register for marking a paper assessment
      description: >
        Primary grades do the monthly assessment in a notebook, so nothing
        reaches the system until a teacher has marked it. Returns the skills
        this class may be marked against and where every student currently
        stands on the one chosen. A student nobody has assessed is present
        with empty values rather than absent.
      parameters:
        - name: classId
          in: query
          required: true
          schema:
            type: integer
        - name: subjectId
          in: query
          required: false
          description: >
            Narrow to one subject the teacher holds in this class. Omitted
            means every subject they hold there, which for a class teacher or
            a primary-grade teacher is every subject the class runs.
          schema:
            type: integer
        - name: skillId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: The register
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentSheet"
        "403":
          description: Not a teacher of that class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /teacher/assessments:
    post:
      operationId: submitAssessment
      tags: [teacher]
      summary: Record a teacher's own marking of a paper assessment
      description: >
        The level may be given without a score. That is the teacher overriding
        the arithmetic, which the requirements ask for by name, and demanding
        a percentage would put a number in the database nobody measured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentSubmission"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitAssessmentResult"
        "400":
          description: A skill or student that does not belong to the class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not a teacher of that class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /teacher/item-analysis:
    get:
      operationId: getItemAnalysis
      tags: [teacher]
      summary: Which questions a class got wrong, worst first
      description: >
        One row per student per question, their most recent answer, so a
        student who retook a quiz four times cannot outvote four students who
        took it once. The most-chosen wrong answer comes with it: a distractor
        that attracts half the class is a specific misunderstanding rather than
        a gap, and the two need different teaching.
      parameters:
        - name: classId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Per-question results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemAnalysis"
        "403":
          description: Not a teacher of that class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /teacher/class-skills:
    get:
      operationId: getClassSkills
      tags: [teacher]
      summary: How a class stands on each skill it has been measured on
      description: >
        Drawn from the answers students have already given, not from a separate
        exam. A skill nobody has attempted is absent rather than reported as
        zero, so the list only ever contains skills there is evidence about.
      parameters:
        - name: classId
          in: query
          required: true
          schema:
            type: integer
        - name: subjectId
          in: query
          required: false
          description: >
            Narrow to one subject the teacher holds in this class. Omitted
            means every subject they hold there, which for a class teacher or
            a primary-grade teacher is every subject the class runs.
          schema:
            type: integer
      responses:
        "200":
          description: Skill standing for the class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassSkills"
        "403":
          description: Not a teacher of that class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/materials:
    get:
      operationId: getAdminMaterials
      tags: [admin]
      summary: Source materials with their file and outline state
      responses:
        "200":
          description: Materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminMaterial"
        "403":
          description: Not an administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/skill-map:
    get:
      operationId: getSkillMap
      tags: [admin]
      summary: Which topics carry which skills, and what is missing
      description: >
        Read-only. Nothing in the product shows this mapping today, yet every
        lesson reaches its book pages through it and every mastery score is
        keyed on the skill it names. An import that lands the topics and the
        skills but drops the links between them leaves a system that can
        timetable lessons and cannot measure anything - and there is currently
        no way to see that has happened short of writing SQL.
      responses:
        "200":
          description: The mapping and its gaps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillMap"
        "403":
          description: Not an administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/skill-chain:
    get:
      operationId: getSkillChain
      tags: [admin]
      summary: Which skill a skill depends on, and whether the chain works
      description: >
        Read-only. The remediation walk follows these links to decide what a
        struggling child should study instead, five levels back at most, and
        until now nothing in the product showed them. Three things about the
        chain are invisible and each breaks it silently: a link that is not
        REQUIRED and APPROVED is never followed, a prerequisite with no
        approved web-ready lesson ends the walk, and a cycle is only stopped by
        the depth limit - the database blocks a skill pointing at itself and
        nothing more.
      responses:
        "200":
          description: The chain, with the links that do nothing marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillChain"
        "403":
          description: Not an administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/materials/{materialId}/file/{filename}:
    post:
      operationId: uploadMaterialFile
      tags: [admin]
      summary: Upload a textbook PDF as a new version of this material
      description: >
        The bytes are checked, not the filename: a file is accepted because it
        starts with %PDF-. Versions are added rather than replaced, because
        outline rows and lesson alignments were made against the page numbers
        of the file already stored. The page offset is set separately, since it
        is nearly always discovered after looking at the book.
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: integer
        - name: filename
          in: path
          required: true
          description: >
            The original name, recorded as a column. It never becomes part of a
            path on disk - the stored file is named from the material's own
            source_code.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/pdf:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedFile"
        "400":
          description: Not a PDF, empty, or no such material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not an administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/materials/{materialId}/page-offset:
    put:
      operationId: setMaterialPageOffset
      tags: [admin]
      summary: Correct the gap between printed and file page numbers
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageOffsetInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageOffsetInput"
        "400":
          description: No file uploaded for this material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/materials/{materialId}/outline:
    get:
      operationId: getMaterialOutline
      tags: [admin]
      summary: A material's sections and page mapping
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Outline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialOutline"
        "403":
          description: Not an administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No such material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: saveMaterialOutline
      tags: [admin]
      summary: Save the sections and the printed-to-file page offset
      description: >
        Sections are matched on outlineCode: known codes are updated, new ones
        inserted. Nothing is deleted here, because a section can already be
        aligned to content and to lessons, and silently dropping it would
        leave those pointing at nothing.
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialOutlineInput"
      responses:
        "200":
          description: Saved outline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialOutline"
        "400":
          description: A page range outside the file, or a duplicate code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not an administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
...
  schemas:
...
    UploadedFile:
      type: object
      properties:
        materialId:
          type: integer
        versionNo:
          type: integer
        filename:
          type: string
        sizeBytes:
          type: integer
        totalPages:
          description: Counted from the file, approximate. The outline is what page numbers are read from.
          type: ["integer", "null"]
        pageOffset:
          type: integer
      required: [materialId, versionNo, filename, sizeBytes, totalPages, pageOffset]
    PageOffsetInput:
      type: object
      properties:
        materialId:
          type: integer
        pageOffset:
          type: integer
      required: [materialId, pageOffset]
    SkillChainLink:
      type: object
      properties:
        dependencyCode:
          type: string
        skillCode:
          type: string
        skillName:
          type: string
        prerequisiteCode:
          type: string
        prerequisiteName:
          type: string
        subjectName:
          type: string
        relationType:
          description: >
            REQUIRED or RECOMMENDED. Only REQUIRED is walked; a RECOMMENDED
            link is a note for whoever plans a year.
          type: string
        importance:
          description: Recorded, but the remediation walk does not read it.
          type: string
        status:
          type: string
        reason:
          description: Why this prerequisite. The column is NOT NULL, so there is always one.
          type: string
        followed:
          description: >
            True when this link is both APPROVED and REQUIRED - that is, when
            remediation actually walks it. Everything else is documentation.
          type: boolean
        prerequisiteHasLesson:
          description: >
            Whether the prerequisite has an approved, web-ready lesson. Without
            one the walk reaches a dead end and keeps looking further back.
          type: boolean
      required:
        - dependencyCode
        - skillCode
        - skillName
        - prerequisiteCode
        - prerequisiteName
        - subjectName
        - relationType
        - importance
        - status
        - reason
        - followed
        - prerequisiteHasLesson
    SkillChain:
      type: object
      properties:
        links:
          type: array
          items:
            $ref: "#/components/schemas/SkillChainLink"
        cycles:
          description: >
            Loops among the followed links, each given as the skill codes
            around it. The database only forbids a skill pointing at itself, so
            a two-step loop is possible; the walk survives it on the depth
            limit alone and returns nonsense.
          type: array
          items:
            type: array
            items:
              type: string
      required: [links, cycles]
    SkillMapSkill:
      type: object
      properties:
        skillCode:
          type: string
        name:
          type: string
        status:
          description: The skill's own review status. Only APPROVED skills reach a student.
          type: string
        isPrimary:
          description: >
            The one skill a topic is mainly about. It decides which book pages
            open with the lesson, so a topic carrying two of them is a fault
            even though the database permits it.
          type: boolean
        mapStatus:
          description: >
            The link's own review status, separate from the skill's. Anything
            but APPROVED is invisible to students.
          type: string
        lessonCount:
          type: integer
      required: [skillCode, name, status, isPrimary, mapStatus, lessonCount]
    SkillMapNode:
      type: object
      properties:
        contentCode:
          type: string
        name:
          type: string
        levelType:
          description: DOMAIN, UNIT, TOPIC, SUBTOPIC or SEGMENT.
          type: string
        subjectName:
          type: string
        status:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/SkillMapSkill"
      required: [contentCode, name, levelType, subjectName, status, skills]
    SkillMapOrphanSkill:
      type: object
      properties:
        skillCode:
          type: string
        name:
          type: string
        subjectName:
          type: string
        status:
          type: string
        lessonCount:
          description: >
            Lessons written against a skill no topic carries. They can be
            taught, but nothing will find them a page in a book.
          type: integer
      required: [skillCode, name, subjectName, status, lessonCount]
    SkillMap:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/SkillMapNode"
        unmappedSkills:
          type: array
          items:
            $ref: "#/components/schemas/SkillMapOrphanSkill"
      required: [nodes, unmappedSkills]
    AdminMaterial:
      type: object
      properties:
        id:
          type: integer
        sourceCode:
          type: string
        title:
          type: ["string", "null"]
        subjectName:
          type: string
        status:
          type: string
        hasFile:
          type: boolean
        filePages:
          type: ["integer", "null"]
        pageOffset:
          type: integer
        sectionCount:
          type: integer
      required:
        - id
        - sourceCode
        - title
        - subjectName
        - status
        - hasFile
        - filePages
        - pageOffset
        - sectionCount
    OutlineSection:
      type: object
      properties:
        id:
          type: ["integer", "null"]
        outlineCode:
          type: string
        printedNumber:
          type: ["string", "null"]
        title:
          type: string
        pageFrom:
          type: ["integer", "null"]
        pageTo:
          type: ["integer", "null"]
        sequenceNo:
          type: integer
        planningPeriodNo:
          type: ["integer", "null"]
          minimum: 1
        usedByLessons:
          description: How many lessons point at this section. Above zero means editing it moves real work.
          type: integer
      required:
        - id
        - outlineCode
        - printedNumber
        - title
        - pageFrom
        - pageTo
        - sequenceNo
        - planningPeriodNo
        - usedByLessons
    MaterialOutline:
      type: object
      properties:
        materialId:
          type: integer
        title:
          type: ["string", "null"]
        pageOffset:
          type: integer
        filePages:
          type: ["integer", "null"]
        planningPeriodCount:
          description: Number of planning periods for this material; not globally fixed to three.
          type: ["integer", "null"]
          minimum: 1
          maximum: 12
        sections:
          type: array
          items:
            $ref: "#/components/schemas/OutlineSection"
      required: [materialId, title, pageOffset, filePages, planningPeriodCount, sections]
    MaterialOutlineInput:
      type: object
      properties:
        pageOffset:
          type: integer
          minimum: 0
        planningPeriodCount:
          type: ["integer", "null"]
          minimum: 1
          maximum: 12
        sections:
          type: array
          items:
            type: object
            properties:
              outlineCode:
                type: string
                minLength: 1
              printedNumber:
                type: ["string", "null"]
              title:
                type: string
                minLength: 1
              pageFrom:
                type: ["integer", "null"]
              pageTo:
                type: ["integer", "null"]
              sequenceNo:
                type: integer
              planningPeriodNo:
                type: ["integer", "null"]
                minimum: 1
                maximum: 12
            required: [outlineCode, printedNumber, title, pageFrom, pageTo, sequenceNo, planningPeriodNo]
      required: [pageOffset, planningPeriodCount, sections]
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
...
    MasteryStatus:
      type: string
      enum: [MASTERED, DEVELOPING, GAP]
    AssessableSkill:
      type: object
      properties:
        skillId:
          type: integer
        skillCode:
          type: string
        skillName:
          type: string
      required: [skillId, skillCode, skillName]
    RosterEntry:
      type: object
      description: >
        One student's standing on the chosen skill. Every field but the
        identity is null for a student nobody has assessed yet.
      properties:
        studentId:
          type: integer
        studentCode:
          type: string
        studentName:
          type: string
        masteryStatus:
          type: ["string", "null"]
        masteryScore:
          type: ["integer", "null"]
        source:
          description: TEACHER where a person entered it, AUTO where the system computed it.
          type: ["string", "null"]
        assessedBy:
          type: ["string", "null"]
        lastAssessedAt:
          type: ["string", "null"]
      required:
        [studentId, studentCode, studentName, masteryStatus, masteryScore,
         source, assessedBy, lastAssessedAt]
    AssessmentSheet:
      type: object
      properties:
        classId:
          type: integer
        className:
          type: string
        gradeLevel:
          type: integer
        stage:
          type: string
          enum: [PRIMARY, SECONDARY]
        skills:
          type: array
          items:
            $ref: "#/components/schemas/AssessableSkill"
        skillId:
          type: ["integer", "null"]
        students:
          type: array
          items:
            $ref: "#/components/schemas/RosterEntry"
      required: [classId, className, gradeLevel, stage, skills, skillId, students]
    AssessmentEntry:
      type: object
      properties:
        studentId:
          type: integer
        status:
          $ref: "#/components/schemas/MasteryStatus"
        score:
          type: ["integer", "null"]
          minimum: 0
          maximum: 100
      required: [studentId, status, score]
    AssessmentSubmission:
      type: object
      properties:
        classId:
          type: integer
        skillId:
          type: integer
        entries:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AssessmentEntry"
      required: [classId, skillId, entries]
    SubmitAssessmentResult:
      type: object
      properties:
        classId:
          type: integer
        skillId:
          type: integer
        recorded:
          type: integer
      required: [classId, skillId, recorded]
    ItemAnalysisRow:
      type: object
      properties:
        itemId:
          type: integer
        prompt:
          type: string
        skillName:
          type: string
        answered:
          type: integer
        correct:
          type: integer
        percentCorrect:
          type: integer
        commonWrongAnswer:
          description: The wrong option chosen most often, or null where nobody got it wrong.
          type: ["string", "null"]
        commonWrongCount:
          type: integer
      required:
        [itemId, prompt, skillName, answered, correct, percentCorrect,
         commonWrongAnswer, commonWrongCount]
    ItemAnalysis:
      type: object
      properties:
        classId:
          type: integer
        className:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemAnalysisRow"
      required: [classId, className, items]
    WeakStudent:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        score:
          type: integer
      required: [studentId, studentName, score]
    ClassSkill:
      type: object
      description: >
        One skill, and how the measured students in the class stand on it.
        `assessed` counts students with any evidence; the three status counts
        add up to it.
      properties:
        skillId:
          type: integer
        skillCode:
          type: string
        skillName:
          type: string
        gradeLevel:
          type: ["integer", "null"]
        assessed:
          type: integer
        gap:
          type: integer
        developing:
          type: integer
        mastered:
          type: integer
        averageScore:
          type: integer
        weakest:
          description: The lowest-scoring students who have not mastered it, capped for reading.
          type: array
          items:
            $ref: "#/components/schemas/WeakStudent"
        weakestTotal:
          description: How many have not mastered it, which may exceed the names listed.
          type: integer
      required:
        [skillId, skillCode, skillName, gradeLevel, assessed, gap, developing,
         mastered, averageScore, weakest, weakestTotal]
    ClassSkills:
      type: object
      properties:
        classId:
          type: integer
        className:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/ClassSkill"
      required: [classId, className, skills]
...